fix: refactor code

+++ Services/User/IUserService.cs
@@ -8,9 +8,9 @@ public interface IUserService
 {
     Task<List<UserResponseDto>> GetUsersAsync();
 
-    Task<UserResponseDto> GetUserAsync(int id);
+    Task<UserResponseDto> GetUserAsync(long id);
 
-    Task UpdateAsync(int id, UpdateUserDto dto);
+    Task UpdateAsync(long id, UpdateUserDto dto);
 
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
